Refactor check_for_single_element_loop to its own module
